mevbot/core/network.py
"""Network optimization using DPDK"""
import os
import ctypes
import logging
from typing import Dict, Optional, List
from dataclasses import dataclass
import asyncio
import socket
import struct

logger = logging.getLogger(__name__)

@dataclass
class DPDKConfig:
    """DPDK configuration"""
    cores: List[int]  # CPU cores to use
    memory_channels: int  # Number of memory channels
    huge_pages: int  # Number of huge pages (2MB each)
    port_mask: int  # Ports to use (bitmask)
    rx_queues: int  # Number of RX queues per port
    tx_queues: int  # Number of TX queues per port

class DPDKManager:
    """Manages DPDK initialization and operation"""
    
    def __init__(self, config: Dict):
        self.config = config
        self.dpdk_config = DPDKConfig(
            cores=config.get('dpdk_cores', [0, 1]),
            memory_channels=config.get('dpdk_memory_channels', 4),
            huge_pages=config.get('dpdk_huge_pages', 1024),  # 2GB total
            port_mask=config.get('dpdk_port_mask', 0x1),
            rx_queues=config.get('dpdk_rx_queues', 4),
            tx_queues=config.get('dpdk_tx_queues', 4)
        )
        self.initialized = False
        self._load_dpdk_lib()
        
    def _load_dpdk_lib(self):
        """Load DPDK library"""
        try:
            # Load DPDK shared library
            self.dpdk = ctypes.CDLL('librte.so')
            logger.info("DPDK library loaded successfully")
        except Exception as e:
            logger.error(f"Failed to load DPDK library: {e}")
            self.dpdk = None
            
    def initialize(self) -> bool:
        """Initialize DPDK"""
        if self.initialized:
            return True
            
        if not self.dpdk:
            logger.error("DPDK library not loaded")
            return False
            
        try:
            # Configure huge pages
            self._configure_huge_pages()
            
            # Initialize EAL (Environment Abstraction Layer)
            args = [
                "mevbot",
                "-l", ",".join(map(str, self.dpdk_config.cores)),
                "-n", str(self.dpdk_config.memory_channels),
                "--huge-dir", "/dev/hugepages",
                "--proc-type", "auto"
            ]
            ret = self.dpdk.rte_eal_init(len(args), args)
            if ret < 0:
                raise RuntimeError("Failed to initialize EAL")
                
            # Initialize memory pools
            self._initialize_mempools()
            
            # Initialize network ports
            self._initialize_ports()
            
            self.initialized = True
            logger.info("DPDK initialized successfully")
            return True
            
        except Exception as e:
            logger.error(f"Failed to initialize DPDK: {e}")
            return False
            
    def _configure_huge_pages(self):
        """Configure huge pages for DPDK"""
        pages_path = "/sys/kernel/mm/hugepages/hugepages-2048kB/nr_hugepages"
        if os.path.exists(pages_path):
            with open(pages_path, 'w') as f:
                f.write(str(self.dpdk_config.huge_pages))
                
    def _initialize_mempools(self):
        """Initialize memory pools for packet buffers"""
        if not self.dpdk:
            return
            
        # Create mempool for RX
        self.rx_mempool = self.dpdk.rte_pktmbuf_pool_create(
            b"rx_mempool",
            8192,  # Number of elements
            256,   # Cache size
            0,     # Private data size
            2048,  # Data room size
            0      # Socket ID (NUMA)
        )
        if not self.rx_mempool:
            raise RuntimeError("Failed to create RX mempool")
            
        # Create mempool for TX
        self.tx_mempool = self.dpdk.rte_pktmbuf_pool_create(
            b"tx_mempool",
            8192,  # Number of elements
            256,   # Cache size
            0,     # Private data size
            2048,  # Data room size
            0      # Socket ID (NUMA)
        )
        if not self.tx_mempool:
            raise RuntimeError("Failed to create TX mempool")
            
    def _initialize_ports(self):
        """Initialize network ports"""
        if not self.dpdk:
            return
            
        # Get number of available ports
        nb_ports = self.dpdk.rte_eth_dev_count_avail()
        if nb_ports == 0:
            raise RuntimeError("No available Ethernet devices")
            
        # Configure each enabled port
        port_mask = self.dpdk_config.port_mask
        for port_id in range(nb_ports):
            if not (port_mask & (1 << port_id)):
                continue
                
            # Port configuration
            port_conf = self._get_port_conf()
            ret = self.dpdk.rte_eth_dev_configure(
                port_id,
                self.dpdk_config.rx_queues,
                self.dpdk_config.tx_queues,
                ctypes.byref(port_conf)
            )
            if ret < 0:
                raise RuntimeError(f"Failed to configure port {port_id}")
                
            # Setup RX queues
            for q in range(self.dpdk_config.rx_queues):
                ret = self.dpdk.rte_eth_rx_queue_setup(
                    port_id, q, 1024,
                    0,  # Socket ID
                    None,  # RX conf
                    self.rx_mempool
                )
                if ret < 0:
                    raise RuntimeError(f"Failed to setup RX queue {q} on port {port_id}")
                    
            # Setup TX queues
            for q in range(self.dpdk_config.tx_queues):
                ret = self.dpdk.rte_eth_tx_queue_setup(
                    port_id, q, 1024,
                    0,  # Socket ID
                    None  # TX conf
                )
                if ret < 0:
                    raise RuntimeError(f"Failed to setup TX queue {q} on port {port_id}")
                    
            # Start the port
            ret = self.dpdk.rte_eth_dev_start(port_id)
            if ret < 0:
                raise RuntimeError(f"Failed to start port {port_id}")
                
    def _get_port_conf(self):
        """Get port configuration"""
        class RteEthConf(ctypes.Structure):
            _fields_ = [
                ("rx_adv_conf", ctypes.c_uint64),
                ("tx_adv_conf", ctypes.c_uint64),
                ("lpbk_mode", ctypes.c_uint32)
            ]
        return RteEthConf()
